web-scraper/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
attrs==25.3.0
azure-core==1.32.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.0
beautifulsoup4==4.12.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
lxml==5.1.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
numpy==2.2.4
pandas==2.2.2
propcache==0.3.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3